Removed flaoting-point calculations

+++ Pathfinding/PathPoint.vb
@@ -2,8 +2,8 @@
 
 Public Class PathPoint
     Public Property Point As Point = New Point(0, 0)
-    Public Property StepValue As Single = 0.0
-    Public Property EstimationValue As Single = 0.0
+    Public Property StepValue As Integer = 0
+    Public Property EstimationValue As Integer = 0
 
     ''' <summary>
     ''' (used to prevent rotations while selecting a path)
